Refactor StartupModal to replace carousel styling with modal card body style; update import paths for step items and remove unused carousel index.
