+++ common/client/components/stepItems/ConfirmItem.vue
